Crate jf_cap

API library for the configurable asset policy system which allows for issuance, transfers and (un)freezing of arbitrary asset that are fully private, yet publicly verifiable and viewable. It is strongly recommended to refer to at least the first two chapters of [Spec] for an introduction to core concepts, data structures and main workflows.

Entities and Their Keys

As explained in Chapter 1.2 of [Spec], there are 5 main roles and all their public/private key related materials can be found under the keys module. Common actions involving a key can usually be found as associated methods of the key struct – e.g. for an viewer to decrypt memos for him/her, use keys::ViewerKeyPair::open_transfer_viewing_memo().

3 Types of Transaction and Their Workflows

As explained in Chapter 2.1 of [Spec], we support 3 types of transaction (txn) – transfer, mint and freeze.

While an actual transaction may contain extra information such as metadata (timestamp), consensus related data etc., our APIs center around the core body of a transaction which we call a TransactionNote. To generate or verify a transaction note of a certain type, go to respective module under the same type name – e.g. transfer::TransferNote::generate_native() or freeze::FreezeNote::verify().

General Workflow

The “privacy-preserving” property of our payment system is largely the result of how each transaction demonstrates its validity and how validators/miners verify them. Namely, in contrast to transaction on Ethereum, which put all necessary info of a txn (amount, sender, receiver etc.) in the open to maintain public verifiability, we utilize cryptographic proofs that can mathematically attest to the “truth of any statement” – statement such as “This txn preserves the balance of total amount of inputs and outputs, plus all senders are legitimate owners of their asset records …”.

To generate such a proof, one would need to first settle the exact statement and deterministically derive one proving key and one verifying key from the statement, the former for the prover of the statement (i.e. the txn creator) whereas the latter for the verifier of the statement (i.e. the validators) – this is called preprocessing, and for the validity statement for each txn type, their APIs can be found under proof module.

One more complication steming from the proof system we use is: before preprocess any concrete statement, we need to retrieve a so-called universal parameter which will be needed during subsequent preprocessings. Reason why it’s called “universal” is because this parameter is independent of the concrete statement of a bounded size.

Therefore, the generic workflow or life cycle of a txn note is as follows:

(during system setup) validators and users fetch the universal parameter1.

sender of a txn runs the preprocessing function to get the proving key of the correct txn type, providing the universal parameter.

sender decides fee, then assembles all other necessary secret witnesses necessary to generate the txn note, a list of receiver memos (one for each output) 2 and a signature signed over the memos.

upon receiving txn note, validators verify it.

for public bulletin board maintainers, verify the signature against the list of memos and the note, and only publish the list of memos if verification passed.

receivers scan the bulletin board to try to decrypt newly added memos to find record openings designated to him which enables him to spend them further.

Proving Keys & Verifying Keys

Ideally, users and validators should have a copy of the proving and verifying keys locally as binary files, and load to memory on demand – utility functions to help with this can be found under parameters module.

However, currently, deserializing these parameter files takes longer than reproducing them, thus we recommend only loading the universal parameter via parameters::load_universal_parameter() and then generate proving keys and verifying keys on demand. For example, for transfer txn, invoke proof::transfer::preprocess() inputting the universal parameter you just loaded to memory 3.

Additional note: there’s potentially different proving/verifying keys for the same txn type but with different number of inputs, outputs and depth of the Merkle Tree accumulating record commitments. Thus, you have to be careful to use the correct ones, or else you will get an Error.

Transaction Note Generation: Fee and Fee Change

Each transaction could decide to include a non-negative fee in native asset type. This is achieved by providing a structs::FeeInput and specify the actual amount of fee: Amount to pay whenever you generate a note. Be careful: the value of FeeInput has to be no smaller than fee, and the difference is fee change which would be included in the output with the same owner as the fee payer.

Transaction Note Generation: Assembling All Necessary Witnesses

There are a dozen of secret witness data required to feed into the txn note generation as inputs, all necessary data structures can be found under structs module, and the best place to learn about how to assemble those input parameters is through our integration test examples.

Let us strongly recommend again: please read example workflows in integration test!

Transaction Note Verification: Batch Verification

In addition to verification on a single txn note, we also offers batch verification which has lower amortized cost – this is offered at txn_batch_verify().

Transaction Fee Collection for Validators

At the ledger level, when a validator proposes a block containing a list of transactions, the validator is entitled to claim all the txn fee specified within the block. We have deviced two APIs regarding txn fee collection:

for block proposer: [prepare_txns_fee_record()]

for the rest of validators who upon validating the block, try to credit the txn fee record to rightful block proposer: [derive_txns_fee_record()]

Again, we recommend readers take a look at the integration test to get an idea of this workflow in actual mock code.

in production, this would be a file downloaded from authority source, but in test and for demo, we can generate them via proof::universal_setup() ↩

Publishing the list of receiver memos to a public bulletin board is optional, as long as the receivers have an authenticated way of receiving the record openings of the output record. ↩

But do expect the slow deserialization problem to be resolved in the future by which time those preprocessings are one-time procedule, and normally you only need to load those proving/verifying keys from files using APIs in parameters. ↩

Modules

bench_utils

constants

Constants used across the whole library

errors

Error types related to transaction API

freeze

Generation and verification of freezing transfer notes

keys

Cryptographic key related data structures

mint

Generation and verification of mint notes (a.k.a Asset Issuance)

parameters

Helper functions to create, save to and fetch from files for prover/verifier parameters.

preclude

crate prelude consisting important traits and structs

proof

SNARK proofs data structures, generation and verification API

structs

record-related data structures in transactions

transfer

Generation and verification of user configurable transfer notes

utils

Utility structures and functions needed for internal purposes, writing tests or benchmarks.

Enums

TransactionNote

A transaction note contains a note of possibly various transaction types, including transfer, mint and freeze.

TransactionVerifyingKey

A transaction verifying key contains a proof verification key of possibly various transaction types, including transfer, mint and freeze.

Functions

calculate_fee

Compute amount of claimable transaction fee

derive_txns_fee_records

Derive a list of record commitments corresponding to fees from a list of transaction notes (within a block). The result is a vector of record commitment which may be further aggregated into a single one by the caller. This function is intended to be called by nodes validating blocks.

sign_receiver_memos

Compute signature over a list of receiver memos

txn_batch_verify

Batch verification for transaction

Type Definitions

AccMemberWitness

The proof of membership in an accumulator (Merkle tree) for an asset record

BaseField

type alias for scalar field of the bls curve

CurveParam

type alias for jubjub curve parameter

KeyPair

The key pair of Schnorr signature scheme

MerkleCommitment

Elements needed for a full commitment to MerkleTree

MerkleFrontier

proof sufficient, with commitment, to define sparse MerkleTree

MerkleLeaf

MerkleTree leaf element

MerkleLeafProof

MerkleTree leaf element with associated authenticated path

MerklePath

An authentication path of a ternary Merkle tree.

MerkleTree

Merkle Tree.

NodeValue

Represents the value for a node in the merkle tree

PairingEngine

type alias for pairing parameter

ScalarField

type alias for scalar field of the jubjub curve

Signature

The signature of Schnorr signature scheme

VerKey

The verification key of Schnorr signature scheme

VerifyingKey

plonk