Function jf_cap::derive_txns_fee_records

pub fn derive_txns_fee_records<C: CapConfig>(
    fee_collectors: &[UserPubKey<C>], 
    fees: &[Amount], 
    blinds: &[BlindFactor<C>]
) -> Result<Vec<RecordCommitment<C>>, TxnApiError>

Derive a list of record commitments corresponding to fees from a list of transaction notes (within a block). The result is a vector of record commitment which may be further aggregated into a single one by the caller. This function is intended to be called by nodes validating blocks.

txns - List of verified transaction notes within a block

owner_pk - Public key of the owner of the collected fee, usually the block proposer’s public key

blind - blinding factor of the record commitment