Module jf_cap::preclude

crate prelude consisting important traits and structs

Re-exports

pub use super::calculate_fee;

pub use super::derive_txns_fee_records;

pub use super::sign_receiver_memos;

pub use super::TransactionNote;

pub use super::TransactionVerifyingKey;

pub use crate::constants::*;

pub use crate::errors::*;

pub use crate::freeze::*;

pub use crate::keys::*;

pub use crate::mint::*;

pub use crate::proof::universal_setup_for_staging;

pub use crate::structs::*;

pub use crate::transfer::AuxInfo as TransferNoteAuxInfo;

pub use crate::transfer::TransferNote;

pub use crate::transfer::TransferNoteInput;

Type Definitions

AccMemberWitness

The proof of membership in an accumulator (Merkle tree) for an asset record

BaseField

type alias for scalar field of the bls curve

CurveParam

type alias for jubjub curve parameter

KeyPair

The key pair of Schnorr signature scheme

MerkleCommitment

Elements needed for a full commitment to MerkleTree

MerkleFrontier

proof sufficient, with commitment, to define sparse MerkleTree

MerkleLeaf

MerkleTree leaf element

MerkleLeafProof

MerkleTree leaf element with associated authenticated path

MerklePath

An authentication path of a ternary Merkle tree.

MerkleTree

Merkle Tree.

NodeValue

Represents the value for a node in the merkle tree

PairingEngine

type alias for pairing parameter

ScalarField

type alias for scalar field of the jubjub curve

Signature

The signature of Schnorr signature scheme

VerKey

The verification key of Schnorr signature scheme

VerifyingKey

plonk