Module jf_cap::proof::transfer

This module contains functions and data structures for

Computing the public parameters

Computing a Transfer proof

Verifying a Transfer proof

Structs

TransferProvingKey

Key for computing the proof associated to a Transfer note

TransferPublicInput

Public inputs of a transfer transaction Struct for the public input of a transfer witness

TransferVerifyingKey

Key for verifying a Transfer note Verifying key of Transfer note and its attributes (n_inputs,n_outputs,tree_depth)

Functions

preprocess

One-time preprocess of the Transfer transaction circuit, proving key and verifying key should be reused for proving/verifying future instances of transfer transaction.