Module jf_cap::constants

Constants used across the whole library

Constants

ATTRS_LEN

the number of identity attributes

AUDIT_DATA_LEN

length in scalars of decrypted transfer audit (address.x, address.y, amount, blinding factor, id_attr_1,…,id_attr_{ATTRS_LEN}

BLS_SCALAR_BIT_CAPACITY

The number of bits that can be reliably stored for BaseField. (Should equal SELF::MODULUS_BITS - 1)

BLS_SCALAR_BIT_LEN

Length of a ScalarField

BLS_SCALAR_REPR_BYTE_LEN

Length of a BaseField representation in bytes

DOM_SEP_DOMESTIC_ASSET

Minted Asset Code domain separator

DOM_SEP_FOREIGN_ASSET

External Asset Code domain separator

DUMMY_ASSET_CODE

Dummy asset code, cannot be 0 (default) or 1(native)

MAX_TIMESTAMP_LEN

The upper bound on the time unit difference between expiry and the current timestamp.

NATIVE_ASSET_CODE

Native asset code, cannot be 0 as then code is identical to default code

PER_ATTR_BYTE_CAPACITY

number of byte can each identityAttribute take.

REVEAL_MAP_LEN

Maximum length of a reveal map := (b_0,b_1,b_2, attr_map) where first three bits enables asset tracing, namely revealing (upk, at, \gamma) and the attr_map is a bitmap for all attributes issued in a identity credential.