Module jf_cap::keys

Cryptographic key related data structures

Role	Data Structures

User (incl Asset Issuer, Validators)	UserKeyPair, UserPubKey

Credential Issuer	CredIssuerKeyPair, CredIssuerPubKey

Tracer/Auditor	AuditorKeyPair, AuditorPubKey

Freezer	FreezerKeyPair, FreezerPubKey

Structs

AuditorKeyPair

Key pair for the auditor

AuditorPubKey

Public key for the auditor

CredIssuerKeyPair

Key pair for the credential issuer

CredIssuerPubKey

Public key for the credential issuer

FreezerKeyPair

Key pair for the freezer

FreezerPubKey

Public key for the freezer

UserKeyPair

A key pair for the user who owns and can consume records (spend asset)

UserPubKey

The public key of a UserKeyPair

Type Definitions

UserAddress

Public address for a user to send assets to/from.